requirements.txt
# 智能人脸识别考勤系统 - 依赖包列表

# GUI框架
PyQt5>=5.15.4
PyQt5-sip>=12.9.0
PyQt5-Qt5>=5.15.2
PyQt5-stubs>=5.15.2.0

# 人脸识别
dlib>=19.22.0

# 图像处理
Pillow>=8.3.1
numpy>=1.21.2

# 数据库
mysql-connector-python>=8.0.26

# 数据可视化
matplotlib>=3.4.3

# API服务
Flask>=2.0.1
requests>=2.26.0

# 其他
opencv-python>=4.5.3.56
pandas>=1.3.3
scipy>=1.7.1